// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  STAFF: 'STAFF',
  CHECKIN: 'CHECKIN',
  VIEWER: 'VIEWER',
  ADMIN: 'ADMIN',
  SUPERADMIN: 'SUPERADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EnrollmentStatus = {
  DRAFT: 'DRAFT',
  REGISTERED: 'REGISTERED',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED'
} as const

export type EnrollmentStatus = (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus]


export const ExamRound = {
  MORNING: 'MORNING',
  AFTERNOON: 'AFTERNOON'
} as const

export type ExamRound = (typeof ExamRound)[keyof typeof ExamRound]


export const EnrollmentSourceType = {
  ONSITE_EXCEL: 'ONSITE_EXCEL',
  SIMULATED_EXCEL: 'SIMULATED_EXCEL',
  MANUAL: 'MANUAL',
  API: 'API'
} as const

export type EnrollmentSourceType = (typeof EnrollmentSourceType)[keyof typeof EnrollmentSourceType]


export const CheckInStatus = {
  SUCCESS: 'SUCCESS',
  DUPLICATE: 'DUPLICATE',
  REJECTED: 'REJECTED'
} as const

export type CheckInStatus = (typeof CheckInStatus)[keyof typeof CheckInStatus]


export const ForfeitRequestStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type ForfeitRequestStatus = (typeof ForfeitRequestStatus)[keyof typeof ForfeitRequestStatus]


export const SystemKey = {
  GENERAL: 'GENERAL'
} as const

export type SystemKey = (typeof SystemKey)[keyof typeof SystemKey]